+++ MergeSort/sample.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 public class sample
 {
 	private static Scanner scanner;
-	private static ArrayList<Integer> arrayList;
+	private static ArrayList<Double> arrayList;
 	private static MergeSort mergeSort;
 
 	public static void main(String[] args)
@@ -26,19 +26,19 @@ public class sample
 
 	private static void initializeClassVariables(int length)
 	{
-		arrayList = new ArrayList<Integer>(length);
+		arrayList = new ArrayList<Double>(length);
 		mergeSort = new MergeSort(arrayList);
 	}
 
 	private static void takeUserInput(int length)
 	{
 		int count = 0;
-		int userNum = 0;
+		double userNum = 0;
 		System.out.print("Elements : ");
 
 		while( count < length )
 		{
-			userNum = scanner.nextInt();
+			userNum = scanner.nextDouble();
 			arrayList.add(count,userNum);
 			count = count + 1;
 		}
